use lsmod to capture only top level module info
The modules we immediately depend on are the most important to record
for errors. Other modules deeper in the tree will just begin to clutter
the modules list. Since the modules section isn't as useful as things
like the stacktrace in general this pruning of prior module list
behavior will be ok.