don't use self to refer to client from transport
These are artifacts from the transports refactor that moved them out of
the client and into their own classes. It doesn't really make sense to
rename the variable when the scopes of the two are the same. This change
is particularly useful because it's easy to read the code and see that
`self` is emitting events, but then it isn't immediately apparent why
binding to `client.transport` isn't working.