Fix propagation of options in Raven.context and Raven.captureException.
Options passed to Raven.context or Raven.captureException weren't getting
propagated (there are unit tests for that, but they only test that the options
get passed to TraceKit.report). This means that options like 'tags' or 'extra'
never actually got passed on to Sentry.

The root of the issue was in TraceKit:

- Parameters passed to TraceKit.report() didn't get passed to notification
  handlers (unless the exception was swallowed and window.onerror didn't get
  called).
- Notification handlers for TraceKit.report() didn't get called at all when
  TraceKit.collectWindowErrors was set to false (unless the exception was
  swallowed, window.onerror didn't get called, and a second call to
  TraceKit.report() was made).