don't remove anonymous functions from stack trace
TraceKit's stacktrace regex for Chrome didn't match anonymous functions,
so all anonymous functions in the stacktrace were being removed before
reporting to Sentry.  This caused errors to be attributed to the wrong
files, and was generally extremely confusing.  This change updates the
regex to not *require* a function name be present in the stacktrace.