Use a more thorough check when exporting.
If an element with ID of `module` is in the document, `module.exports` could be
defined wrongly. This adds a check to ensure `module.exports` exists first.
This commit also removes the `exports` branch which wasn't doing anything,
and also is only used in old environments without `module.exports` support,
such as Narwhal and older versions of Ringo, which are no longer in use.