Add Sentry eventID response
Sets the eventId returned from captureException on ExpressJS middleware. This allows us to use the eventId to inform the user of a tracking ID they can use when in contact with Helpdesk. 

Implements behaviour as shown in https://docs.sentry.io/platforms/node/express/
Likely an oversight that was missed when moving from Raven to the new clients?