feat: Support Windows-style path in RewriteFrame iteratee
The previous default iteratee does not work in Windows because it only detects
paths that begin with `/`. This change ensures paths that begin with `C:\\` are
also accounted for.

This implementation requires that the specified `root` uses Unix-style paths.
So this will not work:

    new RewriteFrames({
        root: 'C:\\Program Files\\Apache\\www'
    })

But this will:

    new RewriteFrames({
        root: '/Program Files/Apache/www'
    })

Should this behaviour change or be noted in the official docs?