DO NOT MERGE: setting the Content-Type to '' does trigger a preflight request
feat: Set outgoing request Content-Type to ''

Sentry's Relay server expects envelopes to have
  Content-Type: application/x-sentry-envelope
however, that header would cause CORS preflight requests that we want to
avoid.

If we don't set the Content-Type header, browsers fill it in with
text/plain. We prefer sending an empty string than setting an incorrect
Content-Type.

Relay accepts the empty string and treats the body as an envelope.