ci: Run browserstack tests only if key is present
For PRs from external contributors, browserstack tests can never succeed
because they depend on secret env var keys that Travis do not provide to
untrusted builds.

This change disables running the tests altogether when the keys are not
available.

The browserstack tests are run anyway in master and release branches, no
changes.