feat: Remove normalization of contexts
All of the context types documented in [1] are flat objects.

The new contexts.trace is the only key under contexts that may have
nested objects under contexts.trace.data. That value holds the data for
the Transaction Span and matches spans[].data for invidual Spans. At the
moment one of the data fields is heavily limited, while the other is
ignored by normalization.

[1]: https://develop.sentry.dev/sdk/event-payloads/contexts/

This is the simplest to implement fix to #2646, thus what I'm doing here
perhaps temporarily.

Fixes #2646