feat: Remove normalization of contexts
All of the context types documented in [1] are flat objects.

The more recent addition contexts.trace is the only context type that
may have nested objects under contexts.trace.data. That value holds the
data for the Transaction Span and matches spans[].data for invidual
Spans.

Removing normalization is one solution to the inconsistency in which
data in Transactions is heavily limited, while data in Span is ignored
by normalization.

This is the simplest to implement fix to #2646, thus what we implement
here to evaluate. We expect that no other key under contexts will be
affected negatively by this change.

[1]: https://develop.sentry.dev/sdk/event-payloads/contexts/

Fixes #2646