feat(tracing): Add beforeNavigate option (#2691)
Adds a beforeNavigate option to the Tracing integration. It allows for users of the integration to change the transaction name before a navigation/pageload transaction has been created.

Currently it defaults to setting the transaction name to `location.pathname` (changed from `location.href` because we don't want to include query params).