rfc(volta): Update to use extends
When I was in the different packages I noticed that my node version would change.

As of Volta 0.8.2 you can use 'extends' for monorepo support, which helps pin node versions in all the packages (except where we need to special case it). Might be nice to have all of node pinned