fix: Reimplement timestamp computation
This is a partial improvement to address inconsistencies in browsers and
how they implement performance.now().

It mitigates observing timestamps in the past for pageload transactions,
but gives no guarantees to navigation transactions.

Depending on the platform, the clock used in performance.now() may stop
when the computer goes to sleep, creating an ever increasing skew. This
skew is more likely to manifest in navigation transactions.

Notable Changes

- Do not polyfill/patch performance.timeOrigin to avoid changing
behavior of third parties that may depend on it. Instead, use an
explicit browserPerformanceTimeOrigin where needed.

- Use a timestamp based on the Date API for error events,
breadcrumbs and envelope header. This should fully resolve the problem
of ingesting events with timestamps in the past, as long as the client
wall clock is reasonably in sync.

- Apply an equivalent workaround that we used for React Native to all
JavaScript environments, essentially resetting the timeOrigin used to
compute monotonic timestamps when execution starts.