Add `optional` parameter to AWSServices integration (#3030)
Initially, I thought there's no need to catch the error of
`require('aws-sdk/global')` because this package is always available on
AWS Lambda and even in the `sam local` emulator.

However, in some environments it could be missing. Examples of such
environments is netlify dev or it could be simply a unit-test environment
without any packages.

Fixes #3000.