feat(serverless): Zip serverless dependencies for AWS Lambda
* Implement extension for AWS Lambda for automatic integration.

* AWS Lambda packages cached

The packages to upload to the AWS Lambda layer are zipped. This zip file is cached.

* AWS Lambda build workflow

Created another workflow to build the AWS Lambda layer by running a script when changes something on @sentry/serverless

* Merged AWS workflow to the main one

Removed the AWS Lambda workflow, and added the build to the main workflow.
Also, included the generated zip file to the artifacts.

* Serverless package versions updated

* Deleted publish script

The publish functionality should not be here, deleted.

* Serverless AWS integration docs minor change

Referenced the docs. Also, updated `yarn.lock`.

* Set default value in the parameter

Instead of setting a default value at the beginning of the function, so that the undefined is avoided, it's now set in the parameter.

* Deleted dependency `path-exists`

* Fixed broken promise

Fixed the broken promise, and removed the `path-exists` dependency from `package.json`.

* ref(aws-lambda): WIP build script

* fix(aws-lambda): Correct API function

Using the function available in the same node version.
Also, removed some logs.

* ref: Revert yarn.lock

* ci: Update

* ref: Reduce imports

* ref: Remove dependency in AWS Lambda build script

* ref: Rename AWS Lambda initializer

Be more specific in the initializer script and rename it to `awslambda-auto`.

* doc: Remove optional step

Co-authored-by: Vladimir Kochnev <marshall-lee@evilmartians.com>
Co-authored-by: Daniel Griesser <daniel.griesser.86@gmail.com>