fix(ember): resolve config path from app root
This should aid in resolving the app configuration from the right location.

Previously, it was calling `require` with a relative path from the `index.js` file of the Sentry addon itself. This broke trying to actually load the configuration file in _other_ apps or addons that might include this as a dependency.

The Ember addon API exposes the configuration path for a given app or addon through the `project.configPath()` method, which we can use instead to look up the location of the project's `config/environment.js` file.