fix(nextjs): Catch exceptions when syncing plug-in's version (#3444)
Fixes https://github.com/getsentry/sentry-javascript/issues/3441
Vercel's file system is read-only, so it throws an exception when trying to rewrite the plugin's version. This happens in runtime in every request in serverless environments.

Users are supposed to build the app before deploying it to Vercel (where the file system isn't read-only), since the plugin version must still be sync-ed.