ref(nextjs): Use resolved paths for `require` calls in config code (#3426)
Currently, we use relative paths for the `require` calls in `syncPluginVersionWithNextVersion`. Though this works most of the time, it depends on the particular structure of the user's `node_modules` directory, which is mostly predictable but not guaranteed. This PR therefore switches to using a combination of `require.resolve` and a recursive search up the file hierarchy to find the `package.json` files we need.

Fixes https://github.com/getsentry/sentry-javascript/issues/3424.
Fixes https://github.com/getsentry/sentry-javascript/issues/3430.