ref(react-router): Make RouteProps typing more generic
Previously we copied over the typings for a react-router Route props.
This means though that when the react-router updates it's types,
there is a chance our types will break. To avoid this in the future,
this patch removes the RouteProps typing and uses a plain
`Record<string, any>` type instead.

The props we do leverage from the RouteProps have checks in the code
and there are tests covering that the RouteProps properties exist,
so making the type more generic is not that risky.