Move towards only wrapping `changeState`
As the docstring say, wrapping `changeState` in the Router should be enough to get performance
monitoring on the frontend. The previous approach was based on wrapping all the exposed Router
methods users could use; but these are using the `changeState` underneath (it seems router actions
always call it but not 100% about it, although it _may_ work implementing far less complexity).

The approach and some code has been taken from
https://github.com/getsentry/sentry-javascript/tree/abhi/next-client-names