ref(ember): Improve sdk consistency with other sdk usage
Considering how being able to pass Sentry config into the init function at runtime is now supported, this cleans up the way the Ember sdk works to be more in-line with our other sdk's (like Vue). This should primarily make our general docs more correct for Ember usage as we can shift to always recommending passing config to init (instead of a lot of special cases) as well as keep maintenance of the sdk simple.

- Switches to exporting an 'init' function (still exporting the old function for backwards compatibility)
- Allows passing browserTracingOptions in the Sentry config at runtime as well, since it's not really an addon option and we might need to pass functions like 'beforeNavigate', which Ember config doesn't really support.