chore(dev): Add `es5` aliases for `build:dist` commands in browser (#3600)
Every other package besides browser uses build:es5 for hopefully self-explanatory purposes. 
The browser package uses build:dist to do the same thing. This adds build:es5 as an alias for 
build:dist, so the commands are standardized across packages and so that they can be run 
as a group in lerna or run when looping through packages manually.