ref: Differentiate between browser/other sessions
Previously in c9b3d107180c03d2256278b266b9ca46d9fe0818, the duration
was updated to be undefined when the browser SDK sent sessions to
Sentry. This patch removes that logic and instead adds a browser flag
to the Session class, which indicates if a duration should be included
or not.