fix(nextjs): Use correct abs path for server init (#3649)
This fixes a bug caused by next build being run on a different machine than next start. Because we've been setting the env var SENTRY_SERVER_INIT_PATH (which we do during build) as an absolute, rather than relative, path, changing machines has made that path invalid when it's later used during server startup. This switches to using a relative path, so that it will work regardless of the machine on which it's run.

Separately, it also fixes a bug where the path (absolute or relative) to the server init file was incorrect when using webpack 5. Under webpack 5, the path in the webpack config's output.path variable points to a different location (one directory deeper) than in webpack 4. To compensate for that difference, the config (as created by next) has a different output.filename value (one including ../ at the beginning), so that then combination points to the same spot as under webpack 4. This PR takes advantage of that by including filename in it derivation of the server init file's final location.

Closes #3648

Co-authored-by: Abhijeet Prasad <aprasad@sentry.io>
Co-authored-by: Katie Byers <lobsterkatie@gmail.com>