ref(nextjs): Stop setting redundant `productionBrowserSourceMaps` in config (#3765)
Before this change, `withSentryConfig` both set `productionBrowserSourceMaps` to `true` in the main `nextjs` config and set `devtool` to `source-map` in the webpack config. 

It turns out, though, that the only thing the former change does is cause nextjs to make the latter change itself. [1] (All other references to `productionBrowserSourceMaps` in the nextjs code are merely type annotations or instances of passing the value from one function to another. [2] The linked code is the only place that value is actually used.) Therefore, since we're making the `devtool` config change ourselves, the `productionBrowserSourceMaps` config change was redundant, and could be removed.

(Of the two config changes, we wanted to remove this one and not the `devtool` one, since we want sourcemaps for both server and client builds. `devtool` affects both, whereas `productionBrowserSourceMaps` (as the name implies) only affects the client build.)

[1] https://github.com/vercel/next.js/blob/fa138358e1df7ca6d1bab7da57e18031d5abbf27/packages/next/build/webpack/config/blocks/base.ts#L39-L43

[2] https://github.com/vercel/next.js/search?q=productionBrowserSourceMaps