fix(react): Make Route typing more generic (#3809)
https://github.com/DefinitelyTyped/DefinitelyTyped/pull/51355 merging in
lead to some more robust type inference for React Router params. This
means that for Typescript >= 4.1.5, certain param types break when using
the `withSentryRouting` higher order component.

To fix this issue, we use a generic instead of a React component type
for the wrapped React component. As we have to use this generic instead
of a specific React type, we have to use @ts-ignore to make the JSX
expressions work with typescript.

Fixes https://github.com/getsentry/sentry-javascript/issues/3808