fix(nextjs): Use local copies for all sentry packages when testing on vercel (#3812)
Previously, when testing a project on vercel against a particular branch of the SDK repo, only the `@sentry/nextjs` package was reflecting the branches changes (by dint of the fact that it was listed as a file dependency pointing to the built SDK). All other `@sentry/*` packages were still getting downloaded from `npm`.

This fixes the install script so that it will force all interdependencies between `@sentry/*` packages to be specified by absolute file paths. That way, only the branch versions of all packages will be used.

There are also some small logging improvements and some generalized cleanup.