Change close/flushes to return false on failure
This matches the comment, and currently rejecting a promise will cause an uncaught promise exception which doesn't get caught and the causes the application to crash. Additionally, rejecting a promise with a boolean doesn't cause a stack trace which made this hard to diagnose. If keeping the rejection is preferred, then it should instead reject with an error object, not just a boolean.