fix(core): Stop rejecting in `flush` and `close` when client undefined (#3846)
Prior to this change, if the SDK had no client defined, calling `Sentry.close()` or `Sentry.flush()` would result in a promise rejection. Since this is not an error per se, this changes that rejection to a logged warning and a promise resolution with `false`. It also fleshes out the docstrings on all of the downstream methods (all of which essentially do the same thing - wait for the event queue to be empty, or timeout if it takes too long).

Fixes https://github.com/getsentry/sentry-javascript/issues/3031.