fix(nextjs): Differentiate between webpack 4 and 5 in server builds (#3878)
In order to cut down on the number of unnecessary files which get uploaded  by the Sentry webpack plugin during nextjs builds, we're quite specific about which filepaths to scan. The change which introduced that specificity[1] missed the fact that under webpack 4, `.next/server/chunks` is not generated. This leads the Sentry webpack plugin to throw an error when it tries to upload it, even though nothing is actually amiss.

This PR fixes that, by checking the webpack version before deciding which files to upload with the plugin.

[1] https://github.com/getsentry/sentry-javascript/pull/3845