Add protocol detection for get/request calls without explict protocol
You can call `get`/`request` on `http`/`https` without providing an
explicit protocol in the request options. In this case the protocol is
automatically set based on which module you made the call to.

Previously calls like this would result in breadcrumbs and traces
without a protocol in the URL creating downstream issues for consumers
of Breadcrumbs as the missing protocol results in URL throwing parsing
errors.

We now attempt to get the protocol by either extracting it from the
`agent` passed as part of the request options, or by extracting it from
the `globalAgent` on which ever http module is pass in `this`.