feat(perf): Add experimental option to improve LCP collection (#3879)
This will add an experimental option (options._metricOptions._reportAllChanges) that will set 'getLCP' to always report. Now that we are recording LCP element, it should be sufficient to collection the additional information and let the user pick the appropriate LCP element in the ui to improve accuracy dynamically instead. We'd like to test this out on Sentry first.

Co-authored-by: Abhijeet Prasad <aprasad@sentry.io>