feat(nextjs): Add ability for integration tests to use linked `@sentry/xxxx` packages (#4019)
In the nextjs integration tests, we use file dependencies for all of the packages in the `sentry-javascript` repo, so that the tests test the local (rather than published) version of the SDK. We don't do the same for `@sentry/cli` or `@sentry/webpack-plugin`, though, because they're in a separate repo and we can't predict where the local copy of that repo lives. As a result, we currently can't (in the nextjs integration tests, at least) test any local changes in either package.

This solves that problem by optionally linking to the local copies of those repos as part of the integration test runner script. In order to use this optional linking:

- To link `@sentry/cli`, set `LINKED_CLI_REPO=<abs path of local sentry-cli repo>`.
- To link `@sentry/webpack-plugin`, set the CLI variable above (since `@sentry/cli` is a dependency of `@sentry/webpack-plugin`, we need to link it in the plugin repo also) as well as `LINKED_PLUGIN_REPO=<abs path of local sentry-webpack-plugin repo>`