fix(docs): Fix docstring for `enabled` option (#4052)
There is an SDK option, `enabled` (referring to the SDK as a whole), which we don't document other than including it in the types. In the types docstring, we currently claim that setting it to `false` will remove any SDK overhead, but this isn't entirely true. (See https://github.com/getsentry/sentry-javascript/issues/4051 for more details). Whether or not this is ideal behavior is a larger question, but for the moment, this changes the docstring to match reality.