fix(nextjs): Support custom `distDir` values in default `RewriteFrames` integration (#4017)
This is a follow-up to https://github.com/getsentry/sentry-javascript/pull/3990/, which makes sure sourcemaps get uploaded from the correct spot when using the nextjs `distDir` option[1]. 

To make sure that the filenames  in stacktrace frames match the names of said sourcemaps, we use the `RewriteFrames` integration. Up until now, we've been hardcoding the replacement it should make, based on nextjs's default output directory. This makes it dynamic, by injecting code at build time which adds the relevant value into `global`, where it is accessible at runtime when we're instantiating the integration.

[1] https://nextjs.org/docs/api-reference/next.config.js/setting-a-custom-build-directory