fix(core): Be stricter about `mechanism` values (#4068)
Though we have an existing `Mechanism` type, we haven't actually been using it in `addExceptionMechanism`. As a result, I didn't catch the fact that the `mechanism` data I added in https://github.com/getsentry/sentry-javascript/pull/4046 was malformed, and therefore being ignored by Sentry.

This fixes both of those problems (the not using of the type and the malformed data being sent). Using the correct type also allowed `addExceptionMechanism` to be streamlined quite a bit.

Finally, given that the question of what the different attributes and their values actually mean has come up more than once, I conferred with the team, and added a docstring to the type. Hopefully this will help anyone who needs to add `mechanism` data in the future.