ref(nextjs): Use real functions in client init tests (#4092)
This changes the structure of the `index.client.ts` tests to stop mocking `configureScope` and `@sentry/react`'s `init` functions and instead spy on the real versions. This makes the structure of the client tests better match the structure of the server tests, and is also necessary for an upcoming PR in which the behavior of those functions is tested directly (meaning they have to actually do the thing they normally would).